requirements.txt
cloudant==2.13.0
tweepy==3.8.0
python-dotenv==0.13.0
slackclient==2.5.0
